pyproject.toml
[tool.poetry]
name = "langchain-project"
version = "0.1.0"
description = ""
authors = ["Khizar Ahmad"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.40.1"
langchain = "^0.2.12"
python-dotenv = "^1.0.1"
langchain-openai = "^0.1.20"
langchain-google-genai = "^1.0.8"
google-generativeai = "^0.7.2"
streamlit = "^1.37.1"
langchain-community = "^0.2.11"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
langserve = "^0.2.2"
sse-starlette = "^2.1.3"
ipykernel = "^6.29.5"
bs4 = "^0.0.2"
pypdf = "^4.3.1"
langchain-text-splitters = "^0.2.2"
faiss-cpu = "^1.8.0.post1"
arxiv = "^2.1.3"
wikipedia = "^1.4.0"
langchainhub = "^0.1.21"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
